refactoring - improve naming
